+++ language/src/main/java/i4gl/exceptions/IndexOutOfBoundsException.java
@@ -1,5 +1,7 @@
 package i4gl.exceptions;
 
+import com.oracle.truffle.api.nodes.Node;
+
 /**
  * Exception thrown when user is accessing an array at non-existing index,
  */
@@ -10,8 +12,12 @@ public class IndexOutOfBoundsException extends I4GLRuntimeException {
      */
     private static final long serialVersionUID = -2867470753887611077L;
 
-    public IndexOutOfBoundsException() {
-        super("Index out of bounds");
+    public IndexOutOfBoundsException(int  index) {
+        super("Index (" + index + ") out of bounds.");
+    }
+
+    public IndexOutOfBoundsException(Node node) {
+        super("Index out of bounds at " + node.toString());
     }
 
 }